% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(input_file)
}
\arguments{
\item{input_file}{A path to an excel file}
}
\value{
a list
}
\description{
Read a standard format MS Excel workbook into a list
}
